+++ src/app/app.module.ts
@@ -3,13 +3,9 @@ import {NgModule} from '@angular/core';
 
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
-import {ProductComponent} from './products/components/product/product.component';
-import {ProductsService} from './products/services/products.service';
-import {CartService} from './cart/services/cart.service';
-import {ProductsListComponent} from './products/components/products-list/products-list.component';
-import {CartComponent} from './cart/components/cart/cart.component';
 import {ProductsModule} from './products/products.module';
 import {CartModule} from './cart/cart.module';
+import {CoreModule} from './core/core.module';
 
 @NgModule({
   declarations: [
@@ -19,7 +15,8 @@ import {CartModule} from './cart/cart.module';
     BrowserModule,
     AppRoutingModule,
     ProductsModule,
-    CartModule
+    CartModule,
+    CoreModule
   ],
   providers: [],
   bootstrap: [AppComponent]