+++ src/app/core/models/config-options.model.ts
@@ -1,0 +1,6 @@
+export class ConfigOptions {
+  constructor(public id: string,
+              public login: string,
+              public email: string) {
+  }
+}