+++ src/app/products/products.module.ts
@@ -3,6 +3,7 @@ import {CommonModule} from '@angular/common';
 import {ProductsListComponent} from './components/products-list/products-list.component';
 import {ProductComponent} from './components/product/product.component';
 import {ProductsService} from './services/products.service';
+import {SharedModule} from '../shared/shared.module';
 
 @NgModule({
   declarations: [
@@ -10,7 +11,8 @@ import {ProductsService} from './services/products.service';
     ProductsListComponent
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    SharedModule
   ],
   providers: [
     ProductsService